Check distribution (#70)
* Add check distribution

* Implement checks into orchestrator API

* Add SetNul on agent check assignment, so we can detect orphaned checks

* Add seeded agents

* Update logging

* Add better config parsing, logging and command handling

* Improve logging for scout

* Config validation + fix crash

* Add basic scheduler + checkHash update handling

* Request retrying logic

* Handle config loading errors

* Add todos

* Fetch actual checks

* Add check body to DTO


* Disable old monitor check


* Fix node-cron issues


* Optimize stale agents check


* Add scheduling of real checks + add basic HTTP checker


* Add Checker struct + add ScoutHeaders and pass it into HTTP check + proper exit for scheduler

* Add TODO comments for distribution


* Add metrics server support

* Describe algorithm for rebalancing

* Switch cron library

* Update algorithm docs for reconcilation

* Throw 401 on agent non-existence

* Add restarting when encountering a 401

* Remove completed todo

* Add pretty errors + upgrade exponential backof + update graceful shutdown for requests


* Add panic recovery


* Handle waitgroups within recovery

* Handle blocking sends + handle panic in update checker


* Add todos

* Only update endAt date if it's not already set

* Add better error handling, serialisation and durations

* Cleanup todos

* Resolve PR comments

---------

+++ apps/api/src/modules/orchestrator/active-agents.ts
@@ -1,8 +1,17 @@
-import type { ConnectedAgent } from "@prisma/client";
+import type { Check, CheckAssignment, ConnectedAgent } from "@prisma/client";
 import { prisma } from "@/modules/db";
 import { getId } from "@/utils/id";
 import { getStaleHeartbeatDate } from "@/modules/orchestrator/utils";
 import { logger } from "@/modules/log";
+import { distributeCheck } from "@/modules/orchestrator/distribution";
+
+export async function getActiveAgent(id: string): Promise<ConnectedAgent | null> {
+  return await prisma.connectedAgent.findFirst({
+    where: {
+      id,
+    },
+  });
+}
 
 export async function registerActiveAgent(agentRegistrationId: string): Promise<ConnectedAgent> {
   return await prisma.connectedAgent.create({
@@ -13,8 +22,8 @@ export async function registerActiveAgent(agentRegistrationId: string): Promise<
   });
 }
 
-export async function refreshActiveAgent(id: string): Promise<void> {
-  await prisma.connectedAgent.update({
+export async function refreshActiveAgent(id: string): Promise<boolean> {
+  const result = await prisma.connectedAgent.updateMany({
     where: {
       id,
     },
@@ -22,6 +31,7 @@ export async function refreshActiveAgent(id: string): Promise<void> {
       lastSeenAt: new Date(),
     },
   });
+  return result.count > 0;
 }
 
 export async function removeActiveAgent(id: string): Promise<void> {
@@ -30,6 +40,7 @@ export async function removeActiveAgent(id: string): Promise<void> {
       id,
     },
   });
+  // TODO register checkAssignments to different agent, use `endAt` to ease them over to the new agent
 }
 
 export async function removeStaleActiveAgents(): Promise<void> {
@@ -40,14 +51,53 @@ export async function removeStaleActiveAgents(): Promise<void> {
       },
     },
   });
-  await prisma.connectedAgent.deleteMany({
+  if (staleAgents.length > 0) {
+    await prisma.connectedAgent.deleteMany({
+      where: {
+        id: {
+          in: staleAgents.map(v => v.id),
+        },
+      },
+    });
+    staleAgents.forEach((agent) => {
+      logger.warn(`Removed stale agent: ${agent.id} for ${agent.registrationId}`);
+    });
+  }
+
+  const orphanedChecks = await prisma.checkAssignment.findMany({
     where: {
-      id: {
-        in: staleAgents.map(v => v.id),
+      agentId: null,
+      endAt: {
+        gte: new Date(),
       },
     },
+    include: {
+      check: true,
+    },
   });
-  staleAgents.forEach((agent) => {
-    logger.warn(`Removed stale agent: ${agent.id} for ${agent.registrationId}`);
+  for (const checkAssignment of orphanedChecks) {
+    await prisma.$transaction(async (tx) => {
+      // TODO put it all on a single agent, so it's quicker and cheaper
+      // It will rely on the daily cost rebalance to sort it out later
+      await distributeCheck(tx, checkAssignment.check);
+      await tx.checkAssignment.delete({
+        where: {
+          id: checkAssignment.id,
+        },
+      });
+    });
+  }
+}
+
+export type FullCheckAssignment = CheckAssignment & { check: Check };
+
+export async function findChecksForAgent(agentId: string): Promise<FullCheckAssignment[]> {
+  return await prisma.checkAssignment.findMany({
+    where: {
+      agentId,
+    },
+    include: {
+      check: true,
+    },
   });
 }
